1.comparing list of person. 2.Adding Nth largest number code. 3.Clean up repo after long, removing target.

+++ src/main/java/practice/core/java/oops/TestingOverrding.java
@@ -21,7 +21,7 @@ class Super{
 
 class Sub extends Super{
 	
-	public short add(int i , int j){
+	public int add(int i , int j){
 		return 0 ;
 	}
 }
